minor code clean up / refactoring
non structural, wont break anything
