+++ GDO.sln.DotSettings
@@ -15,6 +15,7 @@
 	<s:String x:Key="/Default/CodeInspection/ExcludedFiles/FilesAndFoldersToSkip2/=03582490_002DFE29_002D4CBB_002DAFF5_002D98A51EC07E59_002Fd_003AScripts_002Fd_003AMaps_002Fd_003Aol3/@EntryIndexedValue">ExplicitlyExcluded</s:String>
 	<s:String x:Key="/Default/CodeInspection/ExcludedFiles/FilesAndFoldersToSkip2/=03582490_002DFE29_002D4CBB_002DAFF5_002D98A51EC07E59_002Fd_003AScripts_002Fd_003AMaps_002Ff_003Aol3cesium_002Ejs/@EntryIndexedValue">ExplicitlyExcluded</s:String>
 	<s:String x:Key="/Default/CodeInspection/ExcludedFiles/FilesAndFoldersToSkip2/=157EAF14_002D7710_002D448E_002D8260_002D6708F0917B7B_002Fd_003AWeb_002Fd_003ABitcoin_002Fd_003Adata/@EntryIndexedValue">ExplicitlyExcluded</s:String>
+	<s:String x:Key="/Default/CodeInspection/ExcludedFiles/FilesAndFoldersToSkip2/=2159FF04_002DA104_002D4830_002DBAB4_002D86F1F0202967_002Fd_003AScripts_002Fd_003AShanghaiMetro_002Fd_003Aol3/@EntryIndexedValue">ExplicitlyExcluded</s:String>
 	<s:String x:Key="/Default/CodeInspection/ExcludedFiles/FilesAndFoldersToSkip2/=38B42F4F_002D045A_002D4A62_002DAA46_002D063EA1BF7A9A_002Fd_003AScripts_002Fd_003ACity_002Fd_003Aol3/@EntryIndexedValue">ExplicitlyExcluded</s:String>
 	<s:String x:Key="/Default/CodeInspection/ExcludedFiles/FilesAndFoldersToSkip2/=4A27F694_002D1D9A_002D420A_002DA3A6_002D62521EAA047E_002Fd_003ABootstrap/@EntryIndexedValue">ExplicitlyExcluded</s:String>
 	<s:String x:Key="/Default/CodeInspection/ExcludedFiles/FilesAndFoldersToSkip2/=4A27F694_002D1D9A_002D420A_002DA3A6_002D62521EAA047E_002Fd_003ACSS_002Ff_003Ajquery_002Dui_002Ecss/@EntryIndexedValue">ExplicitlyExcluded</s:String>
