+++ GDO/Core/Apps/CompositeApp.cs
@@ -2,8 +2,6 @@
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using GDO.Utility;
-using Microsoft.AspNet.SignalR;
-using Microsoft.AspNet.SignalR.Infrastructure;
 
 namespace GDO.Core.Apps
 {
