+++ GDO/Core/Cave.cs
@@ -585,7 +585,6 @@ namespace GDO.Core
 
         public static List<string> LoadAppConfiguration(string appName, string fileName)
         {
-            List<string> configurationList = new List<string>();
             Directory.SetCurrentDirectory(AppDomain.CurrentDomain.BaseDirectory);
             String path = Directory.GetCurrentDirectory() + @"\Configurations\" + appName;  // TODO using server.map path
             if (Directory.Exists(path))
@@ -611,20 +610,19 @@ namespace GDO.Core
                     }
                 }
             }
-            configurationList = Apps[appName].GetConfigurationList();
+            var configurationList = Apps[appName].GetConfigurationList();
             return configurationList;
         }
 
         public static List<string> UnloadAppConfiguration(string appName, string configName)
         {
-            List<string> configurationList = new List<string>();
             if (Apps[appName].Configurations.ContainsKey(configName))
             {
                 AppConfiguration config;
                 Apps[appName].Configurations.TryRemove(configName, out config);
                 Utilities.RemoveJsonFile(configName, "Configurations\\" + appName);
             }
-            configurationList = Apps[appName].GetConfigurationList();
+            var configurationList = Apps[appName].GetConfigurationList();
             return configurationList;
         }
