+++ GDO/Web/Instances.cshtml
@@ -1,6 +1,4 @@
-﻿@using System.Collections.Generic
-@using System.Threading.Tasks
-@using GDO.Core
+﻿@using GDO.Core
 @using GDO.Core.Apps
 @using GDO.Core.Modules
 @{
